not using setState anymore into createProduct
setSate makes the view re-render all components again, it usefull when we want to change visual components as the same time we change data, but in this case we want to just submit some data information about a certain product, we dont want to update any visual component. So its fine to dont use it.